# specs/002-users-seohun-documents/contracts/api-spec.yaml
openapi: 3.0.0
info:
  title: AgentTrust API
  description: XRPL-based AI Agent License Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Development server
paths:
...
  /downloads/{token}:
    get:
      summary: Download AI agent with token
      tags:
        - Downloads
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        200:
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        403:
          description: Invalid or expired token
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT